moved domain/eqs/cceqs to SeQuant/domain/mbpt/models/cc
