+++ CMakeLists.txt
@@ -238,12 +238,12 @@ add_library(SeQuant
         SeQuant/core/parse/rpn.cpp
         SeQuant/core/parse/regex_sequant.hpp
         SeQuant/core/parse/regex_sequant.cpp
-        SeQuant/domain/eqs/cceqs.hpp
-        SeQuant/domain/eqs/cceqs.cpp
-        SeQuant/domain/eval/eval.hpp
-        SeQuant/domain/eval/cache_manager.hpp
         SeQuant/core/util/rand_color.cpp
         SeQuant/core/util/rand_color.hpp
+        SeQuant/domain/mbpt/models/cc.cpp
+        SeQuant/domain/mbpt/models/cc.hpp
+        SeQuant/domain/eval/eval.hpp
+        SeQuant/domain/eval/cache_manager.hpp
         SeQuant/domain/mbpt/convention.cpp
         SeQuant/domain/mbpt/convention.hpp
         SeQuant/domain/mbpt/op.cpp
