+++ examples/stcc/stcc.cpp
@@ -1,7 +1,7 @@
 #include <SeQuant/core/op.hpp>
 #include <SeQuant/core/timer.hpp>
-#include <SeQuant/domain/eqs/cceqs.hpp>
 #include <SeQuant/domain/mbpt/convention.hpp>
+#include <SeQuant/domain/mbpt/models/cc.hpp>
 #include <SeQuant/domain/mbpt/spin.hpp>
 
 #include <clocale>
@@ -68,9 +68,10 @@ int main(int argc, char* argv[]) {
   };
 
   // Spin-orbital coupled cluster
-  auto cc_r = sequant::eqs::cceqvec{NMAX, NMAX}(true, true, true, true, true);
+  auto cc_r = sequant::mbpt::sr::so::cceqvec{NMAX, true}();
   for (auto i = 1; i < cc_r.size(); ++i) {
-    std::cout << "Spin-orbital CC R" << i << " size: " << cc_r[i]->size() << "\n";
+    std::cout << "Spin-orbital CC R" << i << " size: " << cc_r[i]->size()
+              << "\n";
   }
 
   //
