+++ examples/stcc_rigorous/stcc_rigorous.cpp
@@ -1,7 +1,7 @@
 #include <SeQuant/core/op.hpp>
 #include <SeQuant/core/timer.hpp>
-#include <SeQuant/domain/eqs/cceqs.hpp>
 #include <SeQuant/domain/mbpt/convention.hpp>
+#include <SeQuant/domain/mbpt/models/cc.hpp>
 #include <SeQuant/domain/mbpt/spin.hpp>
 
 #include <clocale>
@@ -68,9 +68,10 @@ int main(int argc, char* argv[]) {
   };
 
   // Spin-orbital coupled cluster
-  auto cc_r = sequant::eqs::cceqvec{NMAX, NMAX}(true, true, true, true, true);
+  auto cc_r = sequant::mbpt::sr::so::cceqvec{NMAX, true}();
   for (auto i = 1; i < cc_r.size(); ++i) {
-    std::cout << "Spin-orbital CC R" << i << " size: " << cc_r[i]->size() << "\n";
+    std::cout << "Spin-orbital CC R" << i << " size: " << cc_r[i]->size()
+              << "\n";
   }
 
   //
@@ -110,8 +111,8 @@ int main(int argc, char* argv[]) {
 
   if (NMAX == 2) {
     runtime_assert(cc_st_r.size() == 3)
-        runtime_assert(cc_st_r.at(1)->size() == 26)    // T1
-        runtime_assert(cc_st_r.at(2)->size() == 55)    // T2
+        runtime_assert(cc_st_r.at(1)->size() == 26)  // T1
+        runtime_assert(cc_st_r.at(2)->size() == 55)  // T2
   } else if (NMAX == 3) {
     runtime_assert(cc_st_r.size() == 4)
         runtime_assert(cc_st_r.at(1)->size() == 30)   // T1
